+++ src/Behat/Mink/Driver/SahiDriver.php
@@ -214,14 +214,11 @@ class SahiDriver implements DriverInterface
     {
         $html = $this->evaluateScript('document.getElementsByTagName("html")[0].innerHTML');
 
-        $html   = html_entity_decode($html);
-        $start  = strpos($html, '<!--SAHI_INJECT_START-->');
-        $finish = strpos($html, '<!--SAHI_INJECT_END-->') ;
-
-        if (false !== $start && false !== $finish) {
-            $finish += strlen('<!--SAHI_INJECT_END-->') - $start;
-            $html    = substr_replace($html, '', $start, $finish);
-        }
+        $html = preg_replace(array(
+            '/<\!--SAHI_INJECT_START--\>.*\<\!--SAHI_INJECT_END--\>/s',
+            '/\<script\>\/\*\<\!\[CDATA\[\*\/\/\*----\>\*\/__sahi.*\<\!--SAHI_INJECT_END--\>/s'
+        ), '', $html);
+        $html = html_entity_decode($html);
 
         return "<html>\n$html\n</html>";
     }
@@ -231,13 +228,16 @@ class SahiDriver implements DriverInterface
      */
     public function find($xpath)
     {
-        $elements = array();
+        $previous = libxml_use_internal_errors(true);
         $document = new \DOMDocument('1.0');
         @$document->loadHTML($this->getContent());
-        $domxpath = new \DOMXPath($document);
+        libxml_clear_errors();
+        libxml_use_internal_errors($previous);
 
-        $count = $domxpath->query($xpath)->length;
-        for ($i = 0; $i < $count; $i++) {
+        $domxpath = new \DOMXPath($document);
+        $domcount = $domxpath->query($xpath)->length;
+        $elements = array();
+        for ($i = 0; $i < $domcount; $i++) {
             $elements[] = new NodeElement(sprintf('(%s)[%d]', $xpath, $i + 1), $this->session);
         }