+++ tests/Behat/Mink/Driver/GoutteDriverTest.php
@@ -2,14 +2,15 @@
 
 namespace Tests\Behat\Mink\Driver;
 
-use Behat\Mink\Driver\GoutteDriver;
-
 require_once 'HeadlessDriverTest.php';
 
 class GoutteDriverTest extends HeadlessDriverTest
 {
-    protected static function configureDriver()
+    protected function setUp()
     {
-        return new GoutteDriver();
+        if (!$this->getMink()->hasSession('goutte')) {
+            $this->getMink()->registerSession('goutte', static::initGoutteSession());
+            $this->getMink()->setDefaultSessionName('goutte');
+        }
     }
 }